backend-ai-langchain/requirements.txt
openai
json
os
datetime
python-dotenv~=1.0.0
langchain~=0.0.292

pypdf2~=3.0.1